"Refactored LeftNav and Navbar components, removed unnecessary code from page.tsx, and deleted landing.tsx and page1.tsx files."
