--- src/app/pages/landing.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-export default function Home() {
-  return (
-    <>
-
-
-    </>
-
-  );
-}
-
