--- src/page1.tsx
@@ -1,15 +1,0 @@
-import Image from "next/image";
-import Navbar from "./components/navbar";
-import React from "react";
-
-export default function Home() {
-  return (
-    <>
-
-    <Navbar/>
-    </>
-
-  );
-}
-
-
